+++ StockAssistPlatform/src/main/java/com/help/stockassistplatform/domain/user/dto/request/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package com.help.stockassistplatform.domain.user.dto;
+package com.help.stockassistplatform.domain.user.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
